refactor: extract Connection from lib.rs into connection/ submodules
Split Connection impl into focused files:
- client.rs: connect, close, cancel_token, accessors, internals
- query.rs: query, execute, simple_query, timeout variants
- transaction_impl.rs: begin, commit, rollback, savepoint
- copy_impl.rs: copy_in, copy_out
- notify_impl.rs: listen, unlisten, notify, wait_for_notification
- pipeline_impl.rs: pipeline, execute_pipeline
- prepare.rs: prepare, register_statement, cache_metrics

lib.rs is now re-exports only. Public API unchanged.
